chat/rooms/serializers.py
from rest_framework import serializers

from .models import Room

class RoomSerializer(serializers.Serializer):
    id = serializers.CharField(read_only=True)
    name = serializers.CharField(max_length=150)

    def create(self, validated_data):
        return Room.objects.create(**validated_data)
    
    def update(self, instance, validated_data):
        instance.name = validated_data.get("name", instance.name)
        instance.save()
        return instance