requirements.txt
Django
channels==3.0.5
pillow
djangorestframework